Move DF cache code from JPAWebloggerImpl to WebloggerImpl
