+++ .gitignore
@@ -1334,3 +1334,4 @@
 .svn/pristine/fe/fed54fa930f7892b0bbc83c097bb0c4048c55dba.svn-base
 .svn/pristine/ff/ff04a0d527e8a897cd7e8956cf6e16d83ec484ed.svn-base
 .svn/pristine/ff/ffa7c527bb7ce8c2ffcb4fae1165f9b86d5cb196.svn-base
+app/target/m2e-wtp/web-resources/META-INF/maven/org.apache.roller/roller-webapp/pom.properties
