+++ app/src/main/java/org/apache/roller/weblogger/business/WebloggerImpl.java
@@ -20,7 +20,11 @@ package org.apache.roller.weblogger.business;
 
 import org.apache.roller.weblogger.business.plugins.PluginManager;
 import java.io.IOException;
+import java.util.HashSet;
+import java.util.List;
 import java.util.Properties;
+import java.util.Set;
+
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.roller.planet.business.PlanetManager;
@@ -33,6 +37,9 @@ import org.apache.roller.weblogger.business.search.IndexManager;
 import org.apache.roller.weblogger.business.runnable.ThreadManager;
 import org.apache.roller.weblogger.business.themes.ThemeManager;
 import org.apache.roller.weblogger.config.PingConfig;
+import org.apache.roller.weblogger.pojos.Weblog;
+import org.apache.roller.weblogger.pojos.WeblogEntry;
+import org.apache.roller.weblogger.pojos.WeblogEntryComment;
 
 
 /**
@@ -75,6 +82,7 @@ public abstract class WebloggerImpl implements Weblogger {
     private final String buildTime;
     private final String buildUser;
     
+    private DFCache dfCache;
     
     protected WebloggerImpl(
         AutoPingManager      autoPingManager,
@@ -128,8 +136,56 @@ public abstract class WebloggerImpl implements Weblogger {
         revision = props.getProperty("ro.revision", "UNKNOWN");
         buildTime = props.getProperty("ro.buildTime", "UNKNOWN");
         buildUser = props.getProperty("ro.buildUser", "UNKNOWN");
+        
+        // LAB 2 - Build the DF Cache
+        
+        this.dfCache = new DFCache();
+        List<Weblog> allWeblogs = this.getWeblogManager().getWeblogs(true, true, null, null, 0, Integer.MAX_VALUE);
+        
+        for(Weblog weblog : allWeblogs){
+            try{
+            	String blogName = weblog.getId();
+                List<WeblogEntry> allEntries = weblog.getRecentWeblogEntries(null, 100);
+            	for (WeblogEntry entry : allEntries){
+            		String docId = blogName + entry.getAnchor();
+            		Set<String> wordSet = getWordsFromDocument(entry);
+            		for (String word : wordSet){
+            			this.dfCache.addItem(docId, word);
+            		}
+            	}
+            }
+            catch(Exception e){
+            	System.out.println("Failed to build DF Cache");
+            	System.out.println(e.toString());
+            	System.out.println(e.getMessage());
+            }
+         }
     }
     
+    // LAB 2 - getWordsFromDocument helper function
+    private Set<String> getWordsFromDocument (WeblogEntry entry){
+        
+    	// Create appended string of a weblog entry and all its comments
+        StringBuilder sb = new StringBuilder();
+        sb.append(entry.getText());
+        for(WeblogEntryComment comment : entry.getComments()){
+        	sb.append(" ");
+        	sb.append(comment.getContent());
+        }
+        
+        // Create array of strings by splitting on word boundaries
+    	String[] words = sb.toString().split("\\b|\\W");
+    	
+    	// Create a set of all words in the document
+    	Set<String> set = new HashSet<String>();
+    	for (String word : words){
+    		if (word.isEmpty())
+    			continue; // don't put empty strings on the map
+    		set.add(word);
+    	}
+    	
+    	return set;
+    }
     
     /**
      * 
@@ -411,6 +467,11 @@ public abstract class WebloggerImpl implements Weblogger {
      */
     public String getBuildUser() {
         return buildUser;
+    }
+    
+    // Lab 2 - Getter for the DFCache object
+    public DFCache getDFCache(){
+    	return dfCache;
     }
     
 }
