+++ app/src/main/java/org/apache/roller/weblogger/business/jpa/JPAWebloggerImpl.java
@@ -56,7 +56,6 @@ public class JPAWebloggerImpl extends WebloggerImpl {
     // a persistence utility class
     private final JPAPersistenceStrategy strategy;
     
-    private DFCache dfCache;
     /**
      * Single constructor.
      * @throws org.apache.roller.weblogger.WebloggerException on any error
@@ -107,55 +106,8 @@ public class JPAWebloggerImpl extends WebloggerImpl {
         
         this.strategy = strategy;
         
-        // LAB 2 - Build the DF Cache
-        
-        this.dfCache = new DFCache();
-        List<Weblog> allWeblogs = this.getWeblogManager().getWeblogs(true, true, null, null, 0, Integer.MAX_VALUE);
-        
-        for(Weblog weblog : allWeblogs){
-            try{
-            	String blogName = weblog.getId();
-                List<WeblogEntry> allEntries = weblog.getRecentWeblogEntries(null, 100);
-            	for (WeblogEntry entry : allEntries){
-            		String docId = blogName + entry.getAnchor();
-            		Set<String> wordSet = getWordsFromDocument(entry);
-            		for (String word : wordSet){
-            			this.dfCache.addItem(docId, word);
-            		}
-            	}
-            }
-            catch(Exception e){
-            	System.out.println("Failed to build DF Cache");
-            	System.out.println(e.toString());
-            	System.out.println(e.getMessage());
-            }
-         }
-        
     }
     
-    private Set<String> getWordsFromDocument (WeblogEntry entry){
-        
-    	// Create appended string of a weblog entry and all its comments
-        StringBuilder sb = new StringBuilder();
-        sb.append(entry.getText());
-        for(WeblogEntryComment comment : entry.getComments()){
-        	sb.append(" ");
-        	sb.append(comment.getContent());
-        }
-        
-        // Create array of strings by splitting on word boundaries
-    	String[] words = sb.toString().split("\\b|\\W");
-    	
-    	// Create a set of all words in the document
-    	Set<String> set = new HashSet<String>();
-    	for (String word : words){
-    		if (word.isEmpty())
-    			continue; // don't put empty strings on the map
-    		set.add(word);
-    	}
-    	
-    	return set;
-    }
     
     public void flush() throws WebloggerException {
         this.strategy.flush();
@@ -179,8 +131,6 @@ public class JPAWebloggerImpl extends WebloggerImpl {
         this.strategy.shutdown();
     }
     
-    public DFCache getDFCache(){
-    	return dfCache;
-    }
+    
 
 }
